backend/requirements.txt
# NeuroPath Backend — Python dependencies
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# AI / ML
tensorflow==2.16.1
google-generativeai==0.8.3

# Image processing
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
matplotlib==3.9.2

# Utilities
python-multipart==0.0.12    # required for FastAPI file uploads
pydantic==2.8.2
